+++ tests/test_pgmm.R
@@ -4,23 +4,23 @@ data("EmplUK", package = "plm")
 z1 <- pgmm(log(emp) ~ lag(log(emp), 1:2) + lag(log(wage), 0:1)
            + log(capital) + lag(log(output), 0:1) | lag(log(emp), 2:99),
             data = EmplUK, effect = "twoways", model = "twosteps")
-summary(z1)
+summary(z1, robust = TRUE) # default
 
 
 z1col <- pgmm(log(emp) ~ lag(log(emp), 1:2) + lag(log(wage), 0:1)
            + log(capital) + lag(log(output), 0:1) | lag(log(emp), 2:99),
            data = EmplUK, effect = "twoways", model = "twosteps", collapse = TRUE)
-summary(z1col)
+summary(z1col, robust = TRUE) # default
 
 z1ind <- pgmm(log(emp) ~ lag(log(emp), 1:2) + lag(log(wage), 0:1)
            + log(capital) + lag(log(output), 0:1) | lag(log(emp), 2:99),
             data = EmplUK, effect = "individual", model = "twosteps")
-summary(z1ind)
+summary(z1ind, robust = TRUE) # default
 
 z1indcol <- pgmm(log(emp) ~ lag(log(emp), 1:2) + lag(log(wage), 0:1)
             + log(capital) + lag(log(output), 0:1) | lag(log(emp), 2:99),
             data = EmplUK, effect = "individual", model = "twosteps")
-summary(z1indcol)
+summary(z1indcol, robust = TRUE) # default
 
 
 ## Blundell and Bond (1998) table 4 (cf DPD for OX p.12 col.4)
@@ -40,12 +40,11 @@ z2b <- pgmm(log(emp) ~ lag(log(emp), 1)+ lag(log(wage), 0:1) +
 summary(z2b, robust = TRUE)
 
 
-
 ### further run tests with various argument values
-summary(z1, robust = TRUE)
-summary(z1col, robust = TRUE)
-summary(z1ind, robust = TRUE)
-summary(z1indcol, robust = TRUE)
+summary(z1, robust = FALSE)
+summary(z1col, robust = FALSE)
+summary(z1ind, robust = FALSE)
+summary(z1indcol, robust = FALSE)
 
 summary(z2, robust = FALSE)
 summary(z2b, robust = FALSE)